Progress on wallet extension
This now completes the circle of the three-way communication which looks
like this now:

  page <=> contentscript <=> background -> popup -> page

Shows already it will be hard to understand what interaction will result
in what effect. A better architecture is to hold all the state in
background and let it handle dispatching to the right parties. Where the
contentscript just stays a dumb event relay between page and
extension/wallet.

* move popup into extra panel from browser action popup
* implement three-way communication between
* send conformation from panel to page
* test icon changing